+++ Documentation/api/json-rpc/README.md
@@ -1,0 +1,3 @@
+# JSON-RPC interface
+
+Delve exposes a [JSON-RPC](http://json-rpc.org/) API interface. Here is an (incomplete) [list of language implementations](http://json-rpc.org/wiki/implementations).
