+++ Documentation/installation/README.md
@@ -1,0 +1,7 @@
+# Installation
+
+Directions for installing Delve on all supported platforms is provided here. Please note you *must* have Go 1.5 or higher installed. Also, if using Go 1.5 you must set `GO15VENDOREXPERIMENT=1` before attempting to install.
+
+- [OSX](osx/install.md)
+- [Linux](linux/install.md)
+- [Windows](windows/install.md)
