+++ Documentation/internal/README.md
@@ -1,0 +1,5 @@
+# Internal Documentation
+
+TODO(derekparker)
+
+This directory will hold documentation around the internals of the debugger and how it works.
