+++ Documentation/usage/dlv_trace.md
@@ -1,0 +1,35 @@
+## dlv trace
+
+Compile and begin tracing program.
+
+### Synopsis
+
+
+Trace program execution. Will set a tracepoint on every function matching the provided regular expression and output information when tracepoint is hit.
+
+```
+dlv trace [package] regexp
+```
+
+### Options
+
+```
+  -p, --pid=0: Pid to attach to.
+  -s, --stack=0: Show stack trace with given depth.
+```
+
+### Options inherited from parent commands
+
+```
+      --accept-multiclient[=false]: Allows a headless server to accept multiple client connection. Note that the server API is not reentrant and clients will have to coordinate
+      --build-flags="": Build flags, to be passed to the compiler.
+      --headless[=false]: Run debug server only, in headless mode.
+      --init="": Init file, executed by the terminal client.
+  -l, --listen="localhost:0": Debugging server listen address.
+      --log[=false]: Enable debugging server logging.
+```
+
+### SEE ALSO
+* [dlv](dlv.md)	 - Delve is a debugger for the Go programming language.
+
+###### Auto generated by spf13/cobra on 19-Feb-2016
