+++ README.md
@@ -12,11 +12,8 @@ The Github issue tracker is for **bugs** only. Please use the [developer mailing
 
 ### About Delve
 
-- [Building](https://github.com/derekparker/delve/wiki/Building)
-- [Usage](https://github.com/derekparker/delve/wiki/Usage)
-- [Command reference](https://github.com/derekparker/delve/wiki/Commands)
-- [Installation Tips & Troubleshooting](https://github.com/derekparker/delve/wiki/Tips-&-Troubleshooting)
-- [Upcoming Features](https://github.com/derekparker/delve/wiki/Upcoming-Features)
-- [Contributing](https://github.com/derekparker/delve/blob/master/CONTRIBUTING.md)
+- [Documentation](Documentation)
+- [Building](Documentation/installation)
+- [Contributing](CONTRIBUTING.md)
 
 Delve is a debugger for the Go programming language. The goal of the project is to provide a simple, full featured debugging tool for Go. Delve should be easy to invoke and easy to use. Chances are if you're using a debugger, most likely things aren't going your way. With that in mind, Delve should stay out of your way as much as possible.
